added basic user permissions

+++ src/posts/views.py
@@ -1,4 +1,4 @@
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
@@ -28,16 +28,19 @@ def post_list(request):
 
 
 def post_create(request):
-    form = PostForm(request.POST or None, request.FILES or None)
-    if form.is_valid():
-        instance = form.save(commit=False)
-        instance.save()
-        messages.success(request, "successfully created")
-        return HttpResponseRedirect(instance.get_absolute_url())
-    context = {
-        "form": form,
-    }
-    return render(request, "post_form.html", context)
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	form = PostForm(request.POST or None, request.FILES or None)
+	if form.is_valid():
+		instance = form.save(commit=False)
+		instance.user = request.user
+		instance.save()
+		messages.success(request, "successfully created")
+		return HttpResponseRedirect(instance.get_absolute_url())
+	context = {
+		"form": form,
+	}
+	return render(request, "post_form.html", context)
 
 
 def post_detail(request, id):
@@ -50,25 +53,28 @@ def post_detail(request, id):
 
 
 def post_update(request, id):
-    instance = get_object_or_404(Post, id=id)
-    form = PostForm(request.POST or None,
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	instance = get_object_or_404(Post, id=id)
+	form = PostForm(request.POST or None,
                     request.FILES or None, instance=instance)
-    if form.is_valid():
-        instance = form.save(commit=False)
-        instance.save()
-        messages.success(request, "ITEM SAVED")
-        return HttpResponseRedirect(instance.get_absolute_url())
-
-    context = {
-        "title": instance.title,
-        "instance": instance,
-        "form": form
-    }
-    return render(request, "post_form.html", context)
+	if form.is_valid():
+		instance = form.save(commit=False)
+		instance.save()
+		messages.success(request, "ITEM SAVED")
+		return HttpResponseRedirect(instance.get_absolute_url())
+	context = {
+		"title": instance.title,
+		"instance": instance,
+		"form": form
+	}
+	return render(request, "post_form.html", context)
 
 
 def post_delete(request, id=None):
-    instance = get_object_or_404(Post, id=id)
-    instance.delete()
-    messages.success(request, "Successfully Deleted")
-    return redirect("posts:list")
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	instance = get_object_or_404(Post, id=id)
+	instance.delete()
+	messages.success(request, "Successfully Deleted")
+	return redirect("posts:list")
